#338 [refactor] sqs 관련 dto common 모듈에서 관리하도록 수정
