--- hous-api/src/main/java/hous/api/config/sqs/dto/MessageDto.java
@@ -1,18 +1,0 @@
-package hous.api.config.sqs.dto;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import lombok.experimental.SuperBuilder;
-
-@ToString
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-@SuperBuilder
-public class MessageDto {
-
-	protected String type;
-}
