+++ hous-common/src/main/java/hous/common/dto/sqs/FirebaseDto.java
@@ -1,4 +1,4 @@
-package hous.api.config.sqs.dto;
+package hous.common.dto.sqs;
 
 import hous.common.constant.MessageType;
 import lombok.AccessLevel;
@@ -13,14 +13,18 @@ import lombok.experimental.SuperBuilder;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @SuperBuilder
-public class SlackUserDeleteFeedbackDto extends MessageDto {
+public class FirebaseDto extends MessageDto {
 
-	private String comment;
+	private String fcmToken;
+	private String title;
+	private String body;
 
-	public static SlackUserDeleteFeedbackDto of(String coment) {
-		return SlackUserDeleteFeedbackDto.builder()
-			.type(MessageType.SLACK_USER_DELETE_FEEDBACK)
-			.comment(coment)
+	public static FirebaseDto of(String fcmToken, String title, String body) {
+		return FirebaseDto.builder()
+			.type(MessageType.FIREBASE)
+			.fcmToken(fcmToken)
+			.title(title)
+			.body(body)
 			.build();
 	}
 }
