+++ hous-common/src/main/java/hous/common/dto/sqs/SlackUserDeleteFeedbackDto.java
@@ -1,4 +1,4 @@
-package hous.notification.config.sqs.dto;
+package hous.common.dto.sqs;
 
 import hous.common.constant.MessageType;
 import lombok.AccessLevel;
@@ -13,14 +13,14 @@ import lombok.experimental.SuperBuilder;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @SuperBuilder
-public class FcmTokenResetDto extends MessageDto {
+public class SlackUserDeleteFeedbackDto extends MessageDto {
 
-	private String fcmToken;
+	private String comment;
 
-	public static FcmTokenResetDto of(String fcmToken) {
-		return FcmTokenResetDto.builder()
-			.type(MessageType.FCM_TOKEN_RESET)
-			.fcmToken(fcmToken)
+	public static SlackUserDeleteFeedbackDto of(String coment) {
+		return SlackUserDeleteFeedbackDto.builder()
+			.type(MessageType.SLACK_USER_DELETE_FEEDBACK)
+			.comment(coment)
 			.build();
 	}
 }
