--- hous-notification/src/main/java/hous/notification/config/sqs/dto/FirebaseDto.java
@@ -1,30 +1,0 @@
-package hous.notification.config.sqs.dto;
-
-import hous.common.constant.MessageType;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import lombok.experimental.SuperBuilder;
-
-@ToString
-@Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@SuperBuilder
-public class FirebaseDto extends MessageDto {
-
-	private String fcmToken;
-	private String title;
-	private String body;
-
-	public static FirebaseDto of(String fcmToken, String title, String body) {
-		return FirebaseDto.builder()
-			.type(MessageType.FIREBASE)
-			.fcmToken(fcmToken)
-			.title(title)
-			.body(body)
-			.build();
-	}
-}
