+++ hous-notification/src/main/java/hous/notification/config/sqs/producer/SqsProducer.java
@@ -13,7 +13,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import hous.common.constant.MessageType;
-import hous.notification.config.sqs.dto.MessageDto;
+import hous.common.dto.sqs.MessageDto;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
